microservicio_busqueda/requirements.txt
Django==4.0
djangorestframework==3.13.1
django-cors-headers==3.13.0
requests==2.28.1
python-dotenv==0.20.0